requirements.txt
Pillow
Flask==1.1.1
